docs: move middleware folder position
